+++ src/main/php/Logger.php
@@ -77,20 +77,6 @@ class Logger
     }
 
     /**
-     * adds a log appender to the logger
-     *
-     * A log appender is responsible for writing the log data.
-     *
-     * @param   \stubbles\log\appender\LogAppender  $logAppender
-     * @return  \stubbles\log\appender\LogAppender  the freshly added log appender instance
-     * @deprecated  since 3.0.0, use addAppender() instead, will be removed with 4.0.0
-     */
-    public function addLogAppender(LogAppender $logAppender)
-    {
-        return $this->addAppender($logAppender);
-    }
-
-    /**
      * checks whether logger has any log appenders
      *
      * @return  bool
@@ -98,17 +84,6 @@ class Logger
     public function hasLogAppenders()
     {
         return (count($this->logAppender) > 0);
-    }
-
-    /**
-     * returns a list of log appenders appended to the logger
-     *
-     * @return  \stubbles\log\appender\LogAppender[]
-     * @deprecated  since 3.0.0, will be removed with 4.0.0
-     */
-    public function getLogAppenders()
-    {
-        return $this->logAppender;
     }
 
     /**